+++ Crypto_App_iOS/Flows/Screens/Home/Views/Data.swift
@@ -1,0 +1,19 @@
+//
+//  ListView.swift
+//  Crypto_App_iOS
+//
+//  Created by Zhenya Pavlovskyi on 06.12.2022.
+//
+
+import SwiftUI
+
+    struct Data: Identifiable {
+        
+        let imageName: Image
+        let nameCoin: String
+        let priceCoin: String
+        let id = UUID()
+        
+    }
+
+
