+++ Crypto_App_iOS/Resources/en.lproj/Localizable.strings
@@ -14,6 +14,4 @@
 
 // Mark: HomeView
 
-"bitcoin" = "Bitcoin";
-"litecoin" = "Litecoin";
-"ethereum" = "Ethereum";
+
